+++ src/main/java/org/example/employee_performance_review_api/domain/model/employee/UpdateEmployeeInput.java
@@ -19,9 +19,9 @@ public class UpdateEmployeeInput {
     private Team team;
     private PerformanceRating rating;
     private EmployeeType type;
-    private int yearlySalary;
+    private double yearlySalary;
     private int employmentTime;
-    private List<Integer> yearlyBonuses;
+    private List<Double> yearlyBonuses;
 
     // TODO finish all the values
 }
